# Cache RBS Environment object