@tompng 2023-06-02
Thank you for the pull request.

I'll pend this pull request for two reason

Maybe different responsibility
 - I think cache should be handled by rbs side. `rbs collection` installs to `project_directory/.gem_rbs_colleciton` and creates lock file, it can create a cache file in project directory as well maybe in the future.
 - Currently, rbs is not doing cache, so using/making a separate gem that adds cache functionality to rbs seems better to me.

Will be hard to port to IRB
 - I want it to be simple before porting to IRB. After that, katakata_irb can do some experimental thing like this pull request.