@tk0miya 2023-06-04
Fair enough. I agree this is not the responsibility of katakata_irb. I'll reconsider other way to optimize loading RBS environment.

I'm closing this now. Please feel free to reopen if you'd still like to keep this pending.